Struct stats_alloc::Instrumented

pub struct Instrumented<T: GlobalAlloc, I: Reporter = NeverReport> { /* fields omitted */ }

An instrumenting middleware which keeps track of allocation, deallocation, and reallocation requests to the underlying global allocator.

Methods

impl Instrumented<System>

pub const fn system() -> Self

Provides access to an instrumented instance of the system allocator.

impl Instrumented<System, AlwaysReport>

pub const fn system_always_report() -> Self

Provides access to an instrumented instance of the system allocator.

impl Instrumented<System, IntervalReport>

pub const fn system_report_every(interval: Duration) -> Self

Provides access to an instrumented instance of the system allocator.

impl<T: GlobalAlloc, I: Reporter> Instrumented<T, I>

pub const fn new(inner: T, default_reporter: I) -> Self

TODO