Struct stats_alloc::Summary

pub struct Summary { /* fields omitted */ }

Allocator statistics

Methods

impl Summary

pub const fn new() -> Self

Initializes an empty stats summary

pub fn reset(&mut self)

Zeroes out the data in this summary, making it equivalent to Stats::new()

pub fn total_bytes_allocated(&self) -> usize

Returns the total quantity of bytes that have been allocated according to this statistical snapshot

pub fn outstanding_bytes_allocated(&self) -> isize

Returns the quantity of bytes that are still allocated according to this statistical snapshot